+++ app/lib/shared/versions.dart
@@ -24,10 +24,10 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const _acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2024.11.21',
+  '2024.11.25',
   // Fallback runtime versions.
+  '2024.11.21',
   '2024.11.18',
-  '2024.11.11',
 ];
 
 /// Sets the current runtime versions.
@@ -62,8 +62,8 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '3.5.0';
-final String toolStableDartSdkVersion = '3.5.4';
-final String toolStableFlutterSdkVersion = '3.24.5';
+final String toolStableDartSdkVersion = '3.6.0-334.4.beta';
+final String toolStableFlutterSdkVersion = '3.27.0-0.2.pre';
 
 final semanticToolStableDartSdkVersion =
     Version.parse(toolStableDartSdkVersion);